requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
tabulate==0.9.0
pydantic-settings==2.0.3
matplotlib==3.8.0
pandas==2.1.0
numpy==1.25.2   
pandas>=2.0
psycopg2-binary>=2.9
plotly>=5.15
python-dotenv>=1.0
pydantic-settings>=2.1
openpyxl>=3.1
